Refactor: Break down api/deploy-contract into smaller components for better structure and maintainability
